make a generic admission server
